get reorder working on new page
